refactor(viewer): separate AnnotationViewer from AnnotationsViewer
